Remove runtime dependencies so geordi starts faster
