+++ bin/geordi
@@ -1,5 +1,6 @@
 #!/usr/bin/env ruby
 
 require 'geordi'
+require 'thor'
 
 Geordi::CLI.start
