+++ lib/geordi/commands/launchy_browser.rb
@@ -1,8 +1,6 @@
 desc '', ''
 def launchy_browser(*args)
   require File.join(File.dirname(__FILE__), '../lib/geordi/cuc')
-
-  require 'rubygems'
   require 'launchy'
 
   Geordi::Cucumber.new.restore_env
